+++ ft-plugin/src/main/java/com/ft/plugin/garble/FTExtension.java
@@ -71,6 +71,12 @@ public class FTExtension {
     public List<String> ignorePackages = new ArrayList<>();
 
     /**
+     * Known WebView class names for inheritance checking, e.g.: ['com/example/CustomWebView', 'com/example/BaseWebView']
+     * These classes will be treated as WebView instances regardless of compilation order
+     */
+    public List<String> knownWebViewClasses = new ArrayList<>();
+
+    /**
      * Default is [project]/build/intermediates/merged_native_libs
      */
     public String nativeLibPath = "";
@@ -113,6 +119,8 @@ public class FTExtension {
                 ", env='" + env + '\'' +
                 ", asmVersion='" + asmVersion + '\'' +
                 ", generateSourceMapOnly='" + generateSourceMapOnly + '\'' +
+                ", ignorePackages=" + ignorePackages +
+                ", knownWebViewClasses=" + knownWebViewClasses +
                 '}';
     }
 }