+++ ft-plugin/src/main/java/com/ft/plugin/garble/asm/FTParameters.java
@@ -11,6 +11,9 @@ public abstract class FTParameters implements InstrumentationParameters {
     public abstract ListProperty<String> getIgnorePackages();
 
     @Input
+    public abstract ListProperty<String> getKnownWebViewClasses();
+
+    @Input
     public abstract Property<String> getAsmVersion();
 
     @Input